% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nets.R
\name{filter_net}
\alias{filter_net}
\title{Filter Net}
\usage{
filter_net(net, words)
}
\arguments{
\item{net}{An object of class \code{phrasenet} as returned by \code{\link{phrase_net}}.}

\item{words}{Vector or words to remove edges.}
}
\description{
Filter edges that contain specific words.
}
\examples{
data(reuters)
phrase_net(reuters, text) \%>\% 
  filter_net(c("a", "the"))

}
