% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nets.R
\name{phrase_net}
\alias{phrase_net}
\title{Phrase Net}
\usage{
phrase_net(data, text, connectors = c("to", "in", "at", "and", "of"))
}
\arguments{
\item{data}{A data.frame or \link[tibble]{tibble} containig \code{text}.}

\item{text}{Bare column name containing text.}

\item{connectors}{Words that establish connections between words and 
ultimately form edges.}
}
\description{
Create phrase network.
}
\examples{
data(reuters)
phrase_net(reuters, text)

}
