% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nets.R
\name{phrase_net}
\alias{phrase_net}
\title{Phrase Net}
\usage{
phrase_net(data, connectors = c("to", "in", "at", "and", "of"), ...)
}
\arguments{
\item{data}{A data.frame or \link[tibble]{tibble} containig \code{text}.}

\item{connectors}{Words that establish connections between words and 
ultimately form edges.}

\item{...}{Arguments to pass to method, namely \code{text}.}
}
\description{
Create phrase network.
}
\examples{
data(reuters)
net1 <- phrase_net(reuters, text = text)
net2 <- phrase_net(reuters$text)

identical(net1, net2)

}
