% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nets.R
\name{plot_sigmajs}
\alias{plot_sigmajs}
\title{Plot sigmajs}
\usage{
plot_sigmajs(net)
}
\arguments{
\item{net}{An object of class \code{phrasenet} as returned by \code{\link{phrase_net}}.}
}
\description{
Plot sigmajs
}
\examples{
data(reuters)

\dontrun{
phrase_net(reuters$text[1:10]) \%>\% 
  plot_sigmajs()
}

}
